ml-service/requirements.txt
Flask==3.0.0
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.26.2
flask-cors==4.0.0
